Casting to int32 after ceil

+++ protein_folding/models/first_model_post_migration.py
@@ -133,7 +133,7 @@ def ScoreModel():
 
   # Reduce, Attend, and Upsample.
   sequence_size = tf.shape(concatenated_features)[1]
-  ideal_sequence_size = tf.math.ceil(tf.cast(sequence_size, tf.float32)/10)*10
+  ideal_sequence_size = tf.cast(tf.math.ceil(tf.cast(sequence_size, tf.float32)/10)*10, tf.int32)
   paddings = tf.constant([[0, 0],[0, ideal_sequence_size-sequence_size],[0, 0]])
   padded_features = tf.pad(concatenated_features, paddings)
   reduced_features = tf.keras.layers.SeparableConv1D(
